Re-implement Random.getBytes
