+++ packages/iov-crypto/src/index.ts
@@ -3,6 +3,7 @@ export { EnglishMnemonic } from "./englishmnemonic";
 export * from "./hmac";
 export * from "./keccak";
 export * from "./libsodium";
+export { Random } from "./random";
 export * from "./secp256k1";
 export { ExtendedSecp256k1Signature, Secp256k1Signature } from "./secp256k1signature";
 export * from "./sha";
