+++ packages/iov-crypto/types/random.d.ts
@@ -1,0 +1,6 @@
+export declare class Random {
+  /**
+   * Returns `count` cryptographically secure random bytes
+   */
+  static getBytes(count: number): Uint8Array;
+}
